Add executing file from a path which has been read from DB
